[argument_parsing_refactor] Moves argument parsing into separate function into commands_base.py for easier testing and development.
