+++ syncano_cli/commands_base.py
@@ -3,6 +3,8 @@ import argparse
 import os
 from collections import defaultdict
 
+import six
+
 
 def argument(*args, **kwargs):
     def wrapper(f):
@@ -39,3 +41,46 @@ class CommandContainer(type):
     @classmethod
     def _register(cls, klass, name, command):
         cls.commands[klass.namespace][name.lower()] = command
+
+
+def add_commands(subparsers, container):
+    for command_name, command in six.iteritems(container):
+        high_lvl_subparser = subparsers.add_parser(command_name, description=command.__doc__)
+        for args, kwargs in getattr(command, 'arguments', []):
+            high_lvl_subparser.add_argument(*args, **kwargs)
+        high_lvl_subparser.set_defaults(func=command)
+
+
+def parse_arguments(args):
+    ACCOUNT_CONFIG_PATH = os.path.join(os.path.expanduser('~'), '.syncano')
+
+    parser = argparse.ArgumentParser(
+        description='Syncano command line tools.'
+    )
+    parser.add_argument('--file', '-f', default='syncano.yml',
+                        help='Instance configuraion file.')
+    parser.add_argument('--config', default=ACCOUNT_CONFIG_PATH,
+                        help='Account configuration file.')
+    parser.add_argument('--key', default=os.environ.get('SYNCANO_API_KEY', None),
+                        help='override ACCOUNT_KEY used for authentication.')
+
+    subparsers = parser.add_subparsers(
+        title='commands',
+        description='valid commands'
+    )
+
+    # add toplevel commands;
+    add_commands(subparsers=subparsers, container=CommandContainer.commands['toplevel'])
+
+    for sub_name, functions in six.iteritems(CommandContainer.commands):
+        if sub_name == 'toplevel':
+            continue
+
+        first_lvl_commands = subparsers.add_parser(sub_name)
+        first_lvl_commands_subparsers = first_lvl_commands.add_subparsers(
+            title='subcommand',
+            description='valid subcommands'
+        )
+        add_commands(subparsers=first_lvl_commands_subparsers, container=functions)
+
+    return parser.parse_args(args)
