+++ tests/test_instance_from_env.py
@@ -1,7 +1,9 @@
 import os
+import mock
 import unittest
 
-import mock
+from syncano_cli import main  # noqa
+from syncano_cli.commands_base import parse_arguments
 
 os.environ.pop('SYNCANO_INSTANCE', None)
 
@@ -11,56 +13,34 @@ class TestException(Exception):
 
 
 class TestInstanceFromEnv(unittest.TestCase):
-    def setUp(self):
-        from syncano_cli import main
-        self.main = main
 
     def test_instance_provided(self):
         os.environ.pop('SYNCANO_INSTANCE', None)
         for cmd in ('pull', 'push'):
-            cmd_mock = mock.Mock(side_effect=TestException(cmd))
-            with mock.patch('syncano.connect', cmd_mock):
-                try:
-                    self.main.cli(['--key', 'api_key', 'sync', cmd, 'args_instance'])
-                except TestException:
-                    cmd_mock.assert_called_with(
-                        api_key='api_key', instance_name='args_instance'
-                    )
+            namespace = parse_arguments(['sync', cmd, 'args_instance'])
+            self.assertEqual(namespace.instance, 'args_instance')
 
     def test_instance_not_provided(self):
         os.environ.pop('SYNCANO_INSTANCE', None)
         for cmd in ('pull', 'push'):
-            cmd_mock = mock.Mock(side_effect=TestException(cmd))
-            with mock.patch('syncano.connect', cmd_mock):
-                try:
-                    self.main.cli(['--key', 'api_key', 'sync', cmd])
-                except SystemExit:
-                    pass
-                else:
-                    raise AssertionError('Instance not provided')
+            try:
+                with mock.patch('sys.stderr'):  # mute errors
+                    parse_arguments(['sync', cmd])
+            except SystemExit:
+                pass
+            else:
+                raise AssertionError('Instance not provided')
 
     def test_instance_from_env(self):
         os.environ['SYNCANO_INSTANCE'] = 'environ_instance'
         for cmd in ('pull', 'push'):
-            cmd_mock = mock.Mock(side_effect=TestException(cmd))
-            with mock.patch('syncano.connect', cmd_mock):
-                try:
-                    self.main.cli(['--key', 'api_key', 'sync', cmd])
-                except TestException:
-                    cmd_mock.assert_called_with(
-                        api_key='api_key', instance_name='environ_instance'
-                    )
+            namespace = parse_arguments(['sync', cmd])
+            self.assertEqual(namespace.instance, 'environ_instance')
         del os.environ['SYNCANO_INSTANCE']
 
     def test_instance_from_env_and_provided(self):
         os.environ['SYNCANO_INSTANCE'] = 'environ_instance'
         for cmd in ('pull', 'push'):
-            cmd_mock = mock.Mock(side_effect=TestException(cmd))
-            with mock.patch('syncano.connect', cmd_mock):
-                try:
-                    self.main.cli(['--key', 'api_key', 'sync', cmd, 'args_instance'])
-                except TestException:
-                    cmd_mock.assert_called_with(
-                        api_key='api_key', instance_name='args_instance'
-                    )
+            namespace = parse_arguments(['sync', cmd, 'args_instance'])
+            self.assertEqual(namespace.instance, 'args_instance')
         del os.environ['SYNCANO_INSTANCE']
